Streams are no longer playing
videoUrl has changed to "rtmp://ios.rtvslo.si/simplevideostreaming32"

Also it seems there is a problem with current playing method using playpath.
So I changed this to pass playback URL as slist parameter.

This should work with mp4 streams. It might also work for flv:// streams, but it would need to be verified.